+++ lumigator/python/mzai/backend/backend/api/routes/experiments.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from fastapi import APIRouter, BackgroundTasks, status
+from fastapi import APIRouter, status
 from lumigator_schemas.experiments import (
     ExperimentCreate,
     ExperimentResponse,
@@ -21,9 +21,8 @@ router = APIRouter()
 def create_experiment(
     service: JobServiceDep,
     request: ExperimentCreate,
-    background_tasks: BackgroundTasks,
 ) -> ExperimentResponse:
-    return service.create_job(JobEvalCreate.model_validate(request.model_dump()), background_tasks)
+    return service.create_job(JobEvalCreate.model_validate(request.model_dump()))
 
 
 @router.get("/{experiment_id}")
