+++ lumigator/python/mzai/backend/backend/api/routes/jobs.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 import loguru
 import requests
-from fastapi import APIRouter, BackgroundTasks, HTTPException, status
+from fastapi import APIRouter, HTTPException, status
 from lumigator_schemas.extras import ListingResponse
 from lumigator_schemas.jobs import (
     Job,
@@ -33,14 +33,8 @@ def create_inference_job(
     job_create_request: JobInferenceCreate,
     request: Request,
     response: Response,
-    background_tasks: BackgroundTasks,
 ) -> JobResponse:
-    # The FastAPI BackgroundTasks object is used to run a function in the background.
-    # It is a wrapper arount Starlette's BackgroundTasks object.
-    # A background task should be attached to a response,
-    # and will run only once the response has been sent.
-    # See here: https://www.starlette.io/background/
-    job_response = service.create_job(job_create_request, background_tasks)
+    job_response = service.create_job(job_create_request)
 
     url = request.url_for(get_job.__name__, job_id=job_response.id)
     response.headers[HttpHeaders.LOCATION] = f"{url}"
@@ -54,9 +48,8 @@ def create_evaluation_job(
     job_create_request: JobEvalCreate,
     request: Request,
     response: Response,
-    background_tasks: BackgroundTasks,
 ) -> JobResponse:
-    job_response = service.create_job(job_create_request, background_tasks)
+    job_response = service.create_job(job_create_request)
 
     url = request.url_for(get_job.__name__, job_id=job_response.id)
     response.headers[HttpHeaders.LOCATION] = f"{url}"
