+++ lumigator/python/mzai/backend/backend/services/jobs.py
@@ -1,15 +1,8 @@
-import asyncio
-import csv
-import json
-from collections.abc import Callable
-from io import BytesIO, StringIO
 from pathlib import Path
-from typing import TYPE_CHECKING
 from uuid import UUID
 
 import loguru
-from fastapi import HTTPException, UploadFile, status
-from lumigator_schemas.datasets import DatasetFormat
+from fastapi import HTTPException, status
 from lumigator_schemas.extras import ListingResponse
 from lumigator_schemas.jobs import (
     JobConfig,
@@ -23,7 +16,6 @@ from lumigator_schemas.jobs import (
 )
 from pydantic import BaseModel
 from ray.job_submission import JobSubmissionClient
-from s3fs import S3FileSystem
 
 from backend import config_templates
 from backend.ray_submit.submission import RayJobEntrypoint, submit_ray_job
@@ -31,9 +23,6 @@ from backend.records.jobs import JobRecord
 from backend.repositories.jobs import JobRepository, JobResultRepository
 from backend.services.datasets import DatasetService
 from backend.settings import settings
-
-if TYPE_CHECKING:
-    from fastapi import BackgroundTasks
 
 
 class JobService:
@@ -174,87 +163,7 @@ class JobService:
 
         return job_params
 
-    def _add_dataset_to_db(self, job_id: UUID, request: JobInferenceCreate):
-        loguru.logger.info("Adding a new dataset entry to the database...")
-        s3 = S3FileSystem()
-
-        # Get the dataset from the S3 bucket
-        result_key = self._get_results_s3_key(job_id)
-        with s3.open(f"{settings.S3_BUCKET}/{result_key}", "r") as f:
-            results = json.loads(f.read())
-
-        dataset = {
-            k: v
-            for k, v in results.items()
-            if k in ["examples", "ground_truth", request.output_field]
-        }
-
-        # Create a CSV in memory
-        csv_buffer = StringIO()
-        csv_writer = csv.writer(csv_buffer)
-        csv_writer.writerow(dataset.keys())
-        csv_writer.writerows(zip(*dataset.values()))
-
-        # Create a binary file from the CSV, since the upload function expects a binary file
-        bin_data = BytesIO(csv_buffer.getvalue().encode("utf-8"))
-        bin_data_size = len(bin_data.getvalue())
-
-        # Figure out the dataset filename
-        dataset_filename = self.data_service.get_dataset(dataset_id=request.dataset).filename
-        dataset_filename = Path(dataset_filename).stem
-        dataset_filename = f"{dataset_filename}-annotated.csv"
-
-        upload_file = UploadFile(
-            file=bin_data,
-            size=bin_data_size,
-            filename=dataset_filename,
-            headers={"content-type": "text/csv"},
-        )
-        dataset_record = self.data_service.upload_dataset(
-            upload_file,
-            format=DatasetFormat.JOB,
-            run_id=job_id,
-            generated=True,
-            generated_by=results["model"],
-        )
-
-        loguru.logger.info(
-            f"Dataset '{dataset_filename}' with ID '{dataset_record.id}' added to the database."
-        )
-
-    async def _run_after_job(self, job_id: UUID, task: Callable = None, *args):
-        """Watches a submitted job and, when it terminates successfully, runs a given task.
-
-        Inputs:
-        - job_id: the UUID of the job to watch
-        - task: the function to be called after the job completes successfully
-        - args: the arguments to be passed to the function `task()`
-        """
-        job_status = self.ray_client.get_job_status(job_id)
-
-        valid_status = [
-            JobStatus.CREATED.value.lower(),
-            JobStatus.PENDING.value.lower(),
-            JobStatus.RUNNING.value.lower(),
-        ]
-        stop_status = [JobStatus.FAILED.value.lower(), JobStatus.SUCCEEDED.value.lower()]
-
-        while job_status.lower() not in stop_status and job_status.lower() in valid_status:
-            loguru.logger.info(f"Watching {job_id}")
-            await asyncio.sleep(5)
-            job_status = self.ray_client.get_job_status(job_id)
-
-        if job_status.lower() == JobStatus.FAILED.value.lower():
-            loguru.logger.error(f"Job {job_id} failed.")
-
-        if job_status.lower() == JobStatus.SUCCEEDED.value.lower():
-            loguru.logger.info(f"Job {job_id} finished successfully.")
-            if task is not None:
-                task(*args)
-
-    def create_job(
-        self, request: JobEvalCreate | JobInferenceCreate, background_tasks: "BackgroundTasks"
-    ) -> JobResponse:
+    def create_job(self, request: JobEvalCreate | JobInferenceCreate) -> JobResponse:
         """Creates a new evaluation workload to run on Ray and returns the response status."""
         if isinstance(request, JobEvalCreate):
             job_type = JobType.EVALUATION
@@ -322,13 +231,6 @@ class JobService:
         )
         loguru.logger.info("Submitting {job_type} Ray job...")
         submit_ray_job(self.ray_client, entrypoint)
-
-        if job_type == JobType.INFERENCE:
-            # Inference jobs produce a new dataset
-            # Add the dataset to the (local) database
-            background_tasks.add_task(
-                self._run_after_job, record.id, self._add_dataset_to_db, record.id, request
-            )
 
         loguru.logger.info("Getting response...")
         return JobResponse.model_validate(record)
